Modify : JumpAttack
점프어택 관련 코드 안넣은 것 수정.
경고 뜨는 코드들 형변환 및 반환값으로 예외처리 했음.

+++ DNF/AttackCollisionComponent.cpp
@@ -33,7 +33,7 @@ int AttackCollisionComponent::GetZBottom() const
 	return (-_posComp->GetZ()) + _zBottom;
 }
 
-void AttackCollisionComponent::SetAttack(const RECT& size, float top, float bottom, float floatingPower)
+void AttackCollisionComponent::SetAttack(const RECT& size, int top, int bottom, float floatingPower)
 {
 	_collider->SetRectSize(size);
 	_zTop = top;
